api/requirements.txt
fastapi
uvicorn
python-dotenv
tortoise-orm
asyncpg
aiohttp
torch
transformers
sentencepiece
accelerate